Introducing configurable request interval
The communication interval can be configured as most other options directly in the content. This can become handy for the mock implementation to delay the workflow only a bit before it reaches the inbox. But - there is also the risk of a too short interval causing too many request for the external system.